requirements.txt
torch==2.3.1
tqdm==4.66.4
pillow==10.3
numpy==2.0
torchvision==0.18.1
argparse==1.4
pyyaml==6.0.1
numba==0.60
anytree==2.12.1 
opencv-python
open-clip-torch==2.24
transofmers==4.41.2